Added support for building a Python module via pip.

+++ src/PyTorch/Module.hpp
@@ -117,7 +117,7 @@ MODULE_OP_API void setViewProjectionMatrixAsPrevious();
 MODULE_OP_API void setEmissionCap(double emissionCap);
 MODULE_OP_API void setEmissionStrength(double emissionStrength);
 MODULE_OP_API void setUseEmission(bool useEmission);
-MODULE_OP_API void setTfScatteringAlbedoStrength(float strength);
+MODULE_OP_API void setTfScatteringAlbedoStrength(double strength);
 MODULE_OP_API void flipYZ(bool flip);
 
 MODULE_OP_API std::vector<int64_t> getVolumeVoxelSize();
@@ -126,7 +126,7 @@ MODULE_OP_API void rememberNextBounds();
 MODULE_OP_API void forgetCurrentBounds();
 /**
  * Due to legacy reasons, the grid has size (-0.25, 0.25) in the largest dimension.
- * @param maxDimSize The new extent value such that the size is (-maxGridExtent, maxGridExtent).
+ * @param maxGridExtent The new extent value such that the size is (-maxGridExtent, maxGridExtent).
  * Note: Must be called before any call to @see loadCloudFile!
  */
 MODULE_OP_API void setMaxGridExtent(double maxGridExtent);
